+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [Feature] Try auto connect on each app open
 - [Feature] Add ability to upload file from file manager
 - [Feature] Add ability to download file from file manager
+- [Feature] Add ability to edit file from file manager
 - [BUGFIX] Exit from emulate screen
 - [BUGFIX] Lock portrait orientation
 - [BUGFIX] Update card not shown when flipper not connected